Making common file.
